requirements.txt
gevent==1.2.1
gunicorn==19.7.1
pycrypto==2.6.1
selenium==3.3.1
sqlalchemy==1.1.6
Pillow==4.0.0
pytesseract==0.1.6
BeautifulSoup4==4.5.3
werkzeug==0.12.1
pdfkit==0.6.1
Jinja2==2.9.5
pymysql==0.7.10
